refac: Refac logging and remove domain dependencies to infra logging
