+++ source/databricks/settlement_report/settlement_report_job/domain/time_series/read_and_filter.py
@@ -16,24 +16,21 @@ from uuid import UUID
 
 from pyspark.sql import DataFrame, functions as F, Column
 
+from settlement_report_job import logging
 from settlement_report_job.domain.market_role import MarketRole
 from settlement_report_job.domain.repository import WholesaleRepository
 from settlement_report_job.domain.system_operator_filter import (
     filter_time_series_on_charge_owner,
 )
-from settlement_report_job.logger import Logger
-from settlement_report_job.infrastructure import logging_configuration
 from settlement_report_job.wholesale.column_names import DataProductColumnNames
 from settlement_report_job.wholesale.data_values import (
     MeteringPointResolutionDataProductValue,
 )
 
-log = Logger(__name__)
+log = logging.Logger(__name__)
 
 
-@logging_configuration.use_span(
-    "settlement_report_job.time_series_factory.read_and_filter"
-)
+@logging.use_span()
 def read_and_filter_for_wholesale(
     period_start: datetime,
     period_end: datetime,
@@ -73,9 +70,7 @@ def read_and_filter_for_wholesale(
     return time_series_points
 
 
-@logging_configuration.use_span(
-    "settlement_report_job.time_series_factory._read_from_view"
-)
+@logging.use_span()
 def _read_from_view(
     period_start: datetime,
     period_end: datetime,
