+++ source/databricks/settlement_report/settlement_report_job/infrastructure/settlement_report_job_args.py
@@ -19,7 +19,6 @@ from argparse import Namespace
 import configargparse
 from configargparse import argparse
 
-from settlement_report_job.infrastructure import logging_configuration
 from settlement_report_job.infrastructure.args_helper import (
     valid_date,
     valid_energy_supplier_ids,
@@ -28,7 +27,7 @@ from settlement_report_job.infrastructure.calculation_type import CalculationTyp
 from settlement_report_job.infrastructure.paths import (
     get_settlement_reports_output_path,
 )
-from settlement_report_job.logger import Logger
+from settlement_report_job.logging import Logger, logging_configuration
 from settlement_report_job.domain.market_role import MarketRole
 from settlement_report_job.domain.settlement_report_args import SettlementReportArgs
 import settlement_report_job.infrastructure.environment_variables as env_vars
