+++ source/databricks/settlement_report/tests/conftest.py
@@ -233,3 +233,14 @@ def spark(
 
     yield session
     session.stop()
+
+
+@pytest.fixture(autouse=True)
+def configure_dummy_logging() -> None:
+    """Ensure that logging hooks don't fail due to _TRACER_NAME not being set."""
+
+    from settlement_report_job.logging.logging_configuration import configure_logging
+
+    configure_logging(
+        cloud_role_name="any-cloud-role-name", tracer_name="any-tracer-name"
+    )
